requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.7
sqlalchemy>=2.0.0
pandas>=2.0.0
openai>=1.30.0
python-dotenv>=1.0.0
httpx>=0.24.0
